Partial refactoring into optimising and calculating parts
